JPA Repository implmented and Exceptions Modified

+++ src/main/java/dev/aditya/productcatalogservice/Service/FakeStoreProductService.java
@@ -1,7 +1,6 @@
 package dev.aditya.productcatalogservice.Service;
 
 import dev.aditya.productcatalogservice.DTO.FakeStoreDTO;
-import dev.aditya.productcatalogservice.Exception.ProductIdMissingException;
 import dev.aditya.productcatalogservice.Exception.ProductNotFoundException;
 import dev.aditya.productcatalogservice.Model.Category;
 import dev.aditya.productcatalogservice.Model.Product;
@@ -9,16 +8,15 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestClient;
-import org.springframework.web.client.RestTemplate;
 
 import java.util.Arrays;
 import java.util.List;
 
 //From this service class we are communicating to another 3rd party service called FakeStore. This acts as a proxy for FakeStore.
 @Service("FakeStoreProductService")
-public class FakeStoreProductService implements ProductServices {
+public class FakeStoreProductService implements ProductService {
 
-    RestTemplate restTemplate;
+    //RestTemplate restTemplate;
     RestClient restClient;
     private final String baseURL = "https://fakestoreapi.com/products";
 
@@ -32,7 +30,7 @@ public class FakeStoreProductService implements ProductServices {
 
 
 
-//    @Autowired(required = false) (required condition is a Fail-Safe in case there is no bean for rest template it'll try injecting in the other constructor)
+//    @Autowired(required = false) (required = false condition is a Fail-Safe in case there is no bean for rest template it'll try injecting in the other constructor)
 //    FakeStoreProductService(RestTemplate restTemplate)
 //    {
 //        this.restTemplate=restTemplate;
@@ -54,7 +52,7 @@ public class FakeStoreProductService implements ProductServices {
 
         if (fakeStoreDTO==null)
         {
-            throw  new ProductNotFoundException("Product doesn't Exist");
+            throw  new ProductNotFoundException();
         }
         return fakeStoreDTO.convertToProduct();
     }
@@ -62,8 +60,7 @@ public class FakeStoreProductService implements ProductServices {
 
 
     @Override
-    public List<Product> getAllProducts()
-    {
+    public List<Product> getAllProducts() throws ProductNotFoundException {
        FakeStoreDTO[] fakeStoreDTOS = restClient.get()
                                                .uri(baseURL)
                                                .retrieve()
@@ -73,7 +70,7 @@ public class FakeStoreProductService implements ProductServices {
 
         if (fakeStoreDTOS == null || fakeStoreDTOS.length==0)
         {
-            throw new NullPointerException("No Products exist yet");
+            throw new ProductNotFoundException("No Products exist yet");
         }
        return Arrays.stream(fakeStoreDTOS)
                     .map(fakeStoreDTO -> fakeStoreDTO.convertToProduct())
@@ -84,13 +81,13 @@ public class FakeStoreProductService implements ProductServices {
 
 
     @Override
-    public Product createNewProduct(String name, String desc, String imageURL, double price, Category category)
+    public Product createNewProduct(String productName, String desc, String imageURL, double price, String categoryName)
     {
         ResponseEntity<FakeStoreDTO> fakeStoreDTOResponseEntity= restClient
                                                                 .post()
                                                                 .uri(baseURL) // <--- This variable is for the URL template
                                                                 //This is the Http request Body. Always pass the object in body() and not it uri unlike rest template.
-                                                                .body(createProductFromParams(name,desc,imageURL,price,category)// This is used in POST and PUT.
+                                                                .body(createProductFromParams(productName,desc,imageURL,price,categoryName)// This is used in POST and PUT.
                                                                      .convertToFakeStoreDTO())
                                                                 .retrieve()
                                                                 .toEntity(FakeStoreDTO.class);
@@ -107,20 +104,24 @@ public class FakeStoreProductService implements ProductServices {
                 return fakeStoreDTOResponseEntity.getBody().convertToProduct();
             }
             else {
-                throw new InternalError("Couldn't create the object. Internal server issue encountered. Please try again later!");
+                throw new InternalError();
             }
         }
         else
         {
-            throw new NullPointerException("Couldn't create the object");
+            throw new NullPointerException();
         }
     }
 
 
 
     @Override
-    public Product deleteProductById(long prodId) throws ProductNotFoundException{
+    public Boolean deleteProductById(long prodId) throws ProductNotFoundException{
         Product product = getProductById(prodId);
+        if(product==null)
+        {
+            throw new ProductNotFoundException();
+        }
         ResponseEntity<FakeStoreDTO> fakeStoreDTOResponseEntity=restClient.delete()
                                                                           .uri(baseURL+"/"+prodId) //No need to pass id twice, Spring can get confused
                                                                           .retrieve()
@@ -132,20 +133,21 @@ public class FakeStoreProductService implements ProductServices {
         {
             if(fakeStoreDTOResponseEntity.getStatusCode().equals(HttpStatus.OK))
             {
-                return fakeStoreDTOResponseEntity.getBody().convertToProduct();
+                //return fakeStoreDTOResponseEntity.getBody().convertToProduct();
+                return true;
             }
             else
             {
                 if(fakeStoreDTOResponseEntity.getStatusCode().is4xxClientError()) {
-                    throw new ProductNotFoundException("Product doesn't Exist");
+                    throw new ProductNotFoundException();
                 }
                 else {
-                    throw new InternalError("Some internal Error occured");
+                    throw new InternalError();
                 }
             }
         }
         else {
-            throw new NullPointerException("Couldn't Delete any product, Null Pointer encountered");
+            throw new NullPointerException();
         }
 
 
@@ -154,15 +156,15 @@ public class FakeStoreProductService implements ProductServices {
     
 
  //Helper Method for creating a model Object
-    private Product createProductFromParams(String name, String desc, String imageURL, double price, Category category)
+    private Product createProductFromParams(String productName, String desc, String imageURL, double price, String categoryName)
     {
         Product product = new Product();
-        product.setId(100l);
-        product.setName(name);
+        product.setId(100);
+        product.setName(productName);
         product.setDescription(desc);
         product.setImageUrl(imageURL);
         product.setPrice(price);
-        product.setCategory(category);
+        product.setCategory(new Category());
         return product;  
     }
 }
